This commit contains the code for the entire project. This commit contains the code that will be used to calculate the actual BMI and display a result.
Step 1: Its create and declares the ActivityMain binding.

Step 2: Declare and create calculateBMI method.

Step 3: Get the values from the view in the method.

Step 4: Use an IF statement to determine if the value are valid, if not then display an error message to the user.

Step 5: preform the calculation and display the result to the user.